# api-server-go/openapi.yaml
openapi: 3.0.3
info:
  title: API Server
  description: |
    Government Data Exchange API Server for managing consumers, providers, and data access workflows.
    This service handles consumer applications, provider submissions, schema management, and access control.
  version: 1.0.0
  contact:
    name: OpenDIF Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api-server.choreo.dev
    description: Choreo Production Environment
  - url: https://api-server-dev.choreo.dev
    description: Choreo Development Environment
  - url: http://localhost:3000
    description: Local Development Environment
paths:
...
  /debug/db:
    get:
      summary: Database Debug Information
      description: Get detailed database connection and table information for debugging
      operationId: debugDatabase
      tags:
        - Debug
      responses:
        '200':
          description: Database debug information
          content:
            application/json:
              schema:
                type: object
                properties:
                  database_connected:
                    type: boolean
                    example: true
                  consumers_table_exists:
                    type: boolean
                    example: true
                  consumers_count:
                    type: integer
                    example: 0
                  table_structure:
                    type: string
                    example: "Columns: consumer_id (character varying), entity_id (character varying), created_at (timestamp with time zone), updated_at (timestamp with time zone)"
                  select_query_test:
                    type: string
                    example: "SELECT query succeeded"
        '503':
          description: Database connection failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "database ping failed: connection refused"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "failed to check table existence: permission denied"
...